agv_vda5050_simulator/requirements.txt
pygame>=2.5.2
paho-mqtt>=2.1.0
numpy>=1.26.4
Pillow>=10.4.0
